Fix object index for ArrayInfo types
ArrayInfo contains the ObjectID record. For records that contain
ArrayInfo, I was neglecting to register the ObjectID from the
contained object. The end effect is that these referenced arrays would
fail to resolve. Fix that.

Also, I neglected to pass the record stream down to a helper parser;
fix that too.